Used flask blueprint to package up code related to authorizations
