+++ app/auth/routes.py
@@ -1,0 +1,97 @@
+from flask import render_template, redirect, url_for, flash, request
+from flask_login import login_user, logout_user, current_user
+from werkzeug.urls import url_parse
+from app import db
+from app.models import User
+from app.auth import bp
+from app.auth.forms import LoginForm, RegistrationForm, \
+    ResetPasswordRequestForm, ResetPasswordForm
+from app.auth.emails import send_password_reset_email
+
+
+@bp.route('/login', methods=['GET', 'POST'])
+def login():
+    """This function implements the view logic for the login page."""
+
+    if current_user.is_authenticated:
+        return redirect(url_for('index'))
+
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(username=form.username.data).first()
+        if not user or not user.check_password(form.password.data):
+            flash("Incorrect username or password. Please try again.")
+            return redirect(url_for('auth.login'))
+        login_user(user, remember=form.remember_me.data)
+        flash("You have logged in successfully!")
+        next_page = request.args.get('next')
+        if not next_page or url_parse(next_page).netloc != '':
+            return redirect(url_for('index'))
+        return redirect(next_page)
+        
+    return render_template('auth/login.html', title='Login', form=form)
+
+
+@bp.route('/logout')
+def logout():
+    """This function implements the logic to log out users."""
+
+    logout_user()
+    return redirect(url_for('auth.login'))
+
+@bp.route('/register', methods=['GET', 'POST'])
+def register():
+    """This function implements the view logic to register new users."""
+
+    if current_user.is_authenticated:
+        return redirect(url_for('index'))
+
+    form = RegistrationForm()
+    if form.validate_on_submit():
+        user = User(username=form.username.data, email=form.email.data)
+        user.set_password(form.password.data)
+        db.session.add(user)
+        db.session.commit()
+        flash("You have registered successfully!")
+        return redirect(url_for('auth.login'))
+
+    return render_template('auth/register.html', title='Register', form=form)
+
+
+@bp.route('/reset_password_request', methods=['GET', 'POST'])
+def reset_password_request():
+    """This view function handles requests to request password reset emails."""
+
+    if current_user.is_authenticated:
+        return redirect(url_for('index'))
+
+    form = ResetPasswordRequestForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(email=form.email.data).first()
+        if user:
+            send_password_reset_email(user)
+        flash(
+            'Check your email for the instructions to reset your password.')
+        return redirect(url_for('auth.login'))
+
+    return render_template('auth/reset_password_request.html', 
+                           title='Reset Password Request', form=form)
+
+
+@bp.route('/reset_password/<token>', methods=['GET', 'POST'])
+def reset_password(token):
+    """This view function handles requests to reset user passwords."""
+
+    user = User.verify_password_reset_token(token)
+    if not user:
+        return redirect(url_for('index'))
+
+    form = ResetPasswordForm()
+    if form.validate_on_submit():
+        user.set_password(form.password.data)
+        db.session.commit()
+        flash("Your password has been reset successfully!")
+        return redirect(url_for('auth.login'))
+
+    return render_template('auth/reset_password.html', title='Reset Password', 
+                           form=form)
