+++ app/templates/auth/login.html
@@ -13,9 +13,9 @@
 </div>
 <br>
 <p>
-    Not registered yet? Click <a href="{{ url_for('register') }}">here</a> to register!
+    Not registered yet? Click <a href="{{ url_for('auth.register') }}">here</a> to register!
 </p>
 <p>
-    Forgot your password? Click <a href="{{ url_for('reset_password_request') }}">here</a> to reset.
+    Forgot your password? Click <a href="{{ url_for('auth.reset_password_request') }}">here</a> to reset.
 </p>
 {% endblock %}
