+++ app/templates/email/reset_password.html
@@ -3,7 +3,7 @@
 </p>
 <p>
     To reset your password, 
-    <a href="{{ url_for('reset_password', token=token, _external=True) }}">
+    <a href="{{ url_for('auth.reset_password', token=token, _external=True) }}">
         click here
     </a>.
 </p>
@@ -11,7 +11,7 @@
     Alternatively, you can paste the following link in your address bar:
 </p>
 <p>
-    {{ url_for('reset_password', token=token, _external=True) }}
+    {{ url_for('auth.reset_password', token=token, _external=True) }}
 </p>
 <p>
     If you have not requested password reset, please simply ignore this message.
