+++ app/templates/email/reset_password.txt
@@ -2,7 +2,7 @@ Dear {{ user.username }}:
 
 To reset your password, please click on the following link:
 
-{{ url_for('reset_password', token=token, _external=True) }}
+{{ url_for('auth.reset_password', token=token, _external=True) }}
 
 If you have not requested a password reset, please simply ignore this message.
